Update README.md
Description: make a program called guessinggame.sh. This program should continuously ask the user to guess the number of files in the current directory, until they guess the correct number. The user is informed if their guess is too high or too low. Once the user guesses the correct number of files in the current directory they should be congratulated. Make date:sat 05/08/2023 Number of lines in guessinggame.sh: 24

KUDOS!!!